+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This file is used to list changes made in each version of the gem.
 - Replace old Modulotech Dockerfile.prod with Rails one.
 - Update production.rb from ModuloprojectGenerator to fit Rails 8.
 - Rewrite dockeruby.rb to dcr.sh for devcontainers and add dc.sh.
+- Update Sidekiq generator to fit devcontainers.
 
 # 1.5.1